Move statics/index to constants/
