+++ src/background/messageListener/lyraApiMessageListener.ts
@@ -1,7 +1,7 @@
 import StellarSdk from "stellar-sdk";
 
 import { ExternalRequest as Request } from "api/types";
-import { EXTERNAL_SERVICE_TYPES } from "statics";
+import { EXTERNAL_SERVICE_TYPES } from "constants/services";
 import { removeQueryParam } from "helpers/urls";
 import { POPUP_WIDTH } from "popup/constants/dimensions";
 import { Sender, SendResponseInterface } from "../types";
