+++ src/background/messageListener/popupMessageListener.ts
@@ -2,8 +2,14 @@ import { KeyManager, KeyManagerPlugins, KeyType } from "@stellar/wallet-sdk";
 import StellarSdk from "stellar-sdk";
 // @ts-ignore
 import { fromMnemonic, generateMnemonic } from "stellar-hd-wallet";
-import { SERVICE_TYPES, APPLICATION_STATE, SERVER_URL } from "statics";
+
+import { SERVER_URL } from "constants/stellar";
+import { SERVICE_TYPES } from "constants/services";
+import { APPLICATION_STATE } from "constants/applicationState";
+
 import { Response as Request } from "api/types";
+import { Sender, SendResponseInterface } from "background/types";
+
 import { removeQueryParam } from "helpers/urls";
 import { Sender, SendResponseInterface } from "../types";
 import {
