+++ src/constants/applicationState.ts
@@ -1,0 +1,7 @@
+export enum APPLICATION_STATE {
+  APPLICATION_LOADING = "APPLICATION_LOADING",
+  APPLICATION_STARTED = "APPLICATION_STARTED",
+  PASSWORD_CREATED = "PASSWORD_CREATED",
+  MNEMONIC_PHRASE_CONFIRMED = "MNEMONIC_PHRASE_CONFIRMED",
+  MNEMONIC_PHRASE_FAILED = "MNEMONIC_PHRASE_FAILED",
+}
