+++ src/constants/extension.ts
@@ -1,0 +1,1 @@
+export const EXTENSION_ID = "kfcdhcbocembkiohhbellipdmpmfpebl";
