+++ src/constants/stellar.ts
@@ -1,0 +1,1 @@
+export const SERVER_URL = "https://horizon-testnet.stellar.org";
