+++ src/popup/components/Header.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import styled from "styled-components";
 import { useSelector } from "react-redux";
 
-import { APPLICATION_STATE } from "statics";
+import { APPLICATION_STATE } from "constants/applicationState";
 
 import { applicationStateSelector } from "popup/ducks/authServices";
